drf_ecommerce/requirement.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
backports.zoneinfo==0.2.1;python_version<"3.9"
colorama==0.4.6
Django==4.2
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.2.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
